+++ packages/clients/walkerjs/src/__tests__/elblayer.test.ts
@@ -181,9 +181,17 @@ describe('ElbLayer', () => {
     expect(w.elbLayer).toBeDefined();
   });
 
+  test('client version equals package.json version', () => {
+    const packageJsonVersion = require('../../package.json').version;
+
+    walkerjs = webClient();
+    expect(walkerjs.config.client).toStrictEqual(packageJsonVersion);
+  });
+
   test('config update', () => {
     const defaultConfig: WebClient.Config = {
       allowed: true,
+      client: expect.any(String),
       consent: {},
       count: expect.any(Number),
       custom: {},