+++ packages/config/tsup/index.d.ts
@@ -1,0 +1,4 @@
+import { Options, defineConfig } from 'tsup';
+
+declare const config: Options;
+export { config, defineConfig };