+++ packages/destinations/web/api/src/index.ts
@@ -8,11 +8,14 @@ export const destinationWebAPI: Function = {
 
   config: {},
 
-  push(event, config) {
+  push(event, config, mapping) {
     const custom = config.custom;
     if (!custom) return;
 
-    const data = JSON.stringify(event);
+    const data = custom.transform
+      ? custom.transform(event, config, mapping) // Transform event data
+      : JSON.stringify(event);
+
     switch (custom.transport) {
       case 'beacon':
         sendAsBeacon(custom.url, data);
@@ -28,7 +31,7 @@ export const destinationWebAPI: Function = {
   },
 };
 
-function sendAsFetch(url: string, data: string) {
+function sendAsFetch(url: string, data: XMLHttpRequestBodyInit) {
   fetch(url, {
     method: 'POST',
     headers: {
@@ -39,11 +42,11 @@ function sendAsFetch(url: string, data: string) {
   });
 }
 
-function sendAsBeacon(url: string, data: string) {
+function sendAsBeacon(url: string, data: XMLHttpRequestBodyInit) {
   navigator.sendBeacon(url, data);
 }
 
-function sendAsXhr(url: string, data: string) {
+function sendAsXhr(url: string, data: XMLHttpRequestBodyInit) {
   const xhr = new XMLHttpRequest();
   xhr.open('POST', url, true);
   xhr.setRequestHeader('Content-type', 'text/plain; charset=utf-8');