+++ packages/destinations/web/api/src/types/index.ts
@@ -1,3 +1,4 @@
+import { Destination, WalkerOS } from '@elbwalker/types';
 import type { WebDestination } from '@elbwalker/walker.js';
 
 export interface Function
@@ -7,9 +8,16 @@ export type Config = WebDestination.Config<CustomConfig, CustomEventConfig>;
 
 export interface CustomConfig {
   url: string;
+  transform?: Transform;
   transport?: Transport;
 }
 
 export interface CustomEventConfig {}
+
+export type Transform = (
+  event: WalkerOS.Event,
+  config?: Config,
+  mapping?: Destination.EventConfig<CustomEventConfig>,
+) => XMLHttpRequestBodyInit;
 
 export type Transport = 'fetch' | 'beacon' | 'xhr';