+++ packages/clients/walkerjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @elbwalker/walker.js
 
+## 2.0.1
+
+### Patch Changes
+
+- Create a built browser version #300
+- add client to config #301
+
 ## 2.0.0
 
 ### Major Changes