+++ packages/destinations/web/piwikpro/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elbwalker/destination-web-piwikpro
 
+## 1.0.1
+
+### Patch Changes
+
+- Create a built browser version #300
+
 ## 1.0.0
 
 ### Major Changes