+++ packages/tagger/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elbwalker/tagger
 
+## 2.0.1
+
+### Patch Changes
+
+- Create a built browser version #300
+
 ## 2.0.0
 
 ### Major Changes